Week9/Day3/ExerciseXP/requirements.txt
smolagents
huggingface_hub
python-dotenv